Closes #38

+++ yo/app/scripts/filters/byteFilter.js
@@ -63,7 +63,7 @@
                     convertedValue = convertedValue.slice(0, (convertedValue.indexOf("."))+3);
                 }                
                 
-                return parseInt(convertedValue);
+                return parseFloat(convertedValue);
 
             }
